vehicle_registry_service/requirements.txt
fastapi
uvicorn
psycopg2-binary
SQLAlchemy
python-jose[cryptography]
python-multipart
fastapi-cors
pydantic[email]
passlib
argon2-cffi